Validation Refactor (#17005)
This refactors the validation component into a few traits and classes. Mainly to separate out areas of focus so the main logic is easier to read. Separated message formatting, replacement formatting, and validation rules.
